[fix] add time and fix sprite

+++ gui/src/cache/Map.cpp
@@ -157,7 +157,7 @@ void MapT::setFrequency(double frequency)
     this->frequency = frequency;
 }
 
-int MapT::getFrequency()
+double MapT::getFrequency()
 {
     return (this->frequency);
 }
